Add files via upload
This is the ipynb notebook for our project. The link to the dataset is provided within the notebook itself. It is based on pix2pix GAN. The generator architecture used is U-Net and the discriminator architecture is Patch-GAN. For evaluation purposes, we have used SSIM, FID, IS. SSIM score was 0.92.